Move KCL value types into their own module
